add users and prepare first process after init
yooo @bxffour, do a scripty thing for bootstrapping the thing so
we can start to build and check where there are errors we can
start fixing. i want to run but I don't have brain to write any
bash scripts right now so a script where that will create the root
dev and let me set the path to the vmlinux and will boot for me
will be great. Thanks!